We are now fully up-to-date with 3.5.  All development should now be
focused on getting us to compile.  Thanks goes to Laurent Montel for KDE4
porting so far.

Nicolas Goutte has already synched the docs up to branches/KDE/3.5/kdegraphics/doc/kolourpaint:492242(HEAD).

This moves us from 1.4_light2pre+ (2005-07-22 22:25:46 +1000 i.e. between
KDE 3.4 and KDE 3.5) to 1.4.1_relight+ (KDE 3.5.1+).  Notable changes:

   * New icons (Danny Allen, Nuno Pinheiro)

   * Tool Box icon size is 22x22, not 16x16, at screen resolution >= 1024x768

   * CTRL + Mouse Wheel = Zoom

   * While freehand selection scaling, holding Shift maintains aspect ratio

   * Prevent accidental drags in the Colour Palette from pasting text
     containing the colour code
     [also in branches/KDE/3.[34]/, branches/kolourpaint/1.2_kde3/]



svn path=/trunk/KDE/kdegraphics/kolourpaint/; revision=515585

+++ views/kpzoomedview.cpp
@@ -94,8 +94,8 @@ void kpZoomedView::adjustToEnvironment ()
     if (document ())
     {
         // TODO: use zoomedDocWidth() & zoomedDocHeight()?
-        resize (transformDocToViewX (document ()->width ()),
-                transformDocToViewY (document ()->height ()));
+        resize ((int) transformDocToViewX (document ()->width ()),
+                (int) transformDocToViewY (document ()->height ()));
     }
 }
